Split equipment.py into separate modules. Draggables / Receivers modules in gui components package.
