+++ src/gui/guild/inventory_grid.py
@@ -1,5 +1,7 @@
-from pyglet.math import Vec2, Vec3, Vec4, Mat4
 from typing import Generator, NamedTuple, Self
+
+from pyglet.math import Mat4, Vec2, Vec3, Vec4
+
 from src.utils.rectangle import Rectangle
 
 
@@ -77,7 +79,7 @@ class SnapGrid:
         return screen_loc in self.bounds
 
     def locations_in_rows(self) -> Generator[Vec2, None, None]:
-        top_left = GridLoc.snap(Vec2(0, self.grid_bounds.t - 1))
+        top_left = GridLoc.snap(Vec2(0, self.grid_bounds.t))
         along = lambda gl: GridLoc(gl.x + 1, gl.y)
         next_row = lambda gl: GridLoc(0, gl.y - 1)
